I2C transmit without dma is unstable

+++ include/hal/tch_i2c.h
@@ -56,8 +56,8 @@ struct tch_iic_cfg_t {
 
 struct tch_iic_handle_t{
 	tchStatus (*close)(tch_iicHandle* self);
-	tchStatus (*write)(tch_iicHandle* self,uint16_t addr,const void* wb,size_t sz);
-	tchStatus (*read)(tch_iicHandle* self,uint16_t addr,void* rb,size_t sz,uint32_t timeout);
+	tchStatus (*write)(tch_iicHandle* self,uint16_t addr,const void* wb,int32_t sz);
+	tchStatus (*read)(tch_iicHandle* self,uint16_t addr,void* rb,int32_t sz,uint32_t timeout);
 };
